Middleware for Storage
